app/report_service/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# ORM & DB
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10

# AWS SDK
boto3==1.38.32
s3transfer==0.13.0
botocore==1.38.32

# LangChain & Graph
langchain==0.3.25
langchain-core==0.3.64
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-sdk==0.1.70
langchain-aws==0.2.24

# Pydantic & Settings
pydantic==2.11.7
pydantic-settings==2.10.1

# JWT & Auth
python-jose==3.5.0
PyJWT==2.10.1
python-dotenv==1.1.0

# S3 PreSigned + Requesting
requests==2.32.3

# Typing / Utility
typing_extensions==4.14.0

# Logging
colorlog==6.9.0
